# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Taqeem Reports Tool
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ReportStatus = (typeof ReportStatus)[keyof typeof ReportStatus];

export const ReportStatus = {
  pending: "pending",
  extracted: "extracted",
  reviewed: "reviewed",
  submitted: "submitted",
} as const;

export interface Report {
  id: number;
  /** @nullable */
  reportNumber?: string | null;
  /** @nullable */
  reportDate?: string | null;
  /** @nullable */
  valuationDate?: string | null;
  /** @nullable */
  inspectionDate?: string | null;
  /** @nullable */
  commissionDate?: string | null;
  /** @nullable */
  requestNumber?: string | null;
  status: ReportStatus;
  /** @nullable */
  valuerName?: string | null;
  /** @nullable */
  valuerPercentage?: number | null;
  /** @nullable */
  licenseNumber?: string | null;
  /** @nullable */
  licenseDate?: string | null;
  /** @nullable */
  membershipNumber?: string | null;
  /** @nullable */
  membershipType?: string | null;
  /** @nullable */
  secondValuerName?: string | null;
  /** @nullable */
  secondValuerPercentage?: number | null;
  /** @nullable */
  secondValuerLicenseNumber?: string | null;
  /** @nullable */
  secondValuerMembershipNumber?: string | null;
  /** @nullable */
  taqeemReportNumber?: string | null;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  intendedUser?: string | null;
  /** @nullable */
  reportType?: string | null;
  /** @nullable */
  valuationPurpose?: string | null;
  /** @nullable */
  valuationBasis?: string | null;
  /** @nullable */
  propertyType?: string | null;
  /** @nullable */
  propertySubType?: string | null;
  /** @nullable */
  region?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  district?: string | null;
  /** @nullable */
  street?: string | null;
  /** @nullable */
  blockNumber?: string | null;
  /** @nullable */
  plotNumber?: string | null;
  /** @nullable */
  planNumber?: string | null;
  /** @nullable */
  propertyUse?: string | null;
  /** @nullable */
  deedNumber?: string | null;
  /** @nullable */
  deedDate?: string | null;
  /** @nullable */
  ownerName?: string | null;
  /** @nullable */
  ownershipType?: string | null;
  /** @nullable */
  buildingPermitNumber?: string | null;
  /** @nullable */
  buildingStatus?: string | null;
  /** @nullable */
  buildingAge?: string | null;
  /** @nullable */
  landArea?: number | null;
  /** @nullable */
  buildingArea?: number | null;
  /** @nullable */
  basementArea?: number | null;
  /** @nullable */
  annexArea?: number | null;
  /** @nullable */
  floorsCount?: number | null;
  /** @nullable */
  coordinates?: string | null;
  /** @nullable */
  valuationMethod?: string | null;
  /** @nullable */
  marketValue?: number | null;
  /** @nullable */
  incomeValue?: number | null;
  /** @nullable */
  costValue?: number | null;
  /** @nullable */
  finalValue?: number | null;
  /** @nullable */
  pricePerMeter?: number | null;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  commercialRegNumber?: string | null;
  /** @nullable */
  pdfFileName?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateReportBodyStatus =
  (typeof CreateReportBodyStatus)[keyof typeof CreateReportBodyStatus];

export const CreateReportBodyStatus = {
  pending: "pending",
  extracted: "extracted",
  reviewed: "reviewed",
  submitted: "submitted",
} as const;

export interface CreateReportBody {
  /** @nullable */
  reportNumber?: string | null;
  status?: CreateReportBodyStatus;
}

export interface UpdateReportBody {
  /** @nullable */
  reportNumber?: string | null;
  /** @nullable */
  reportDate?: string | null;
  /** @nullable */
  valuationDate?: string | null;
  /** @nullable */
  inspectionDate?: string | null;
  /** @nullable */
  commissionDate?: string | null;
  /** @nullable */
  requestNumber?: string | null;
  /** @nullable */
  valuerName?: string | null;
  /** @nullable */
  valuerPercentage?: number | null;
  /** @nullable */
  licenseNumber?: string | null;
  /** @nullable */
  licenseDate?: string | null;
  /** @nullable */
  membershipNumber?: string | null;
  /** @nullable */
  membershipType?: string | null;
  /** @nullable */
  secondValuerName?: string | null;
  /** @nullable */
  secondValuerPercentage?: number | null;
  /** @nullable */
  secondValuerLicenseNumber?: string | null;
  /** @nullable */
  secondValuerMembershipNumber?: string | null;
  /** @nullable */
  taqeemReportNumber?: string | null;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  intendedUser?: string | null;
  /** @nullable */
  reportType?: string | null;
  /** @nullable */
  valuationPurpose?: string | null;
  /** @nullable */
  valuationBasis?: string | null;
  /** @nullable */
  propertyType?: string | null;
  /** @nullable */
  propertySubType?: string | null;
  /** @nullable */
  region?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  district?: string | null;
  /** @nullable */
  street?: string | null;
  /** @nullable */
  blockNumber?: string | null;
  /** @nullable */
  plotNumber?: string | null;
  /** @nullable */
  planNumber?: string | null;
  /** @nullable */
  propertyUse?: string | null;
  /** @nullable */
  deedNumber?: string | null;
  /** @nullable */
  deedDate?: string | null;
  /** @nullable */
  ownerName?: string | null;
  /** @nullable */
  ownershipType?: string | null;
  /** @nullable */
  buildingPermitNumber?: string | null;
  /** @nullable */
  buildingStatus?: string | null;
  /** @nullable */
  buildingAge?: string | null;
  /** @nullable */
  landArea?: number | null;
  /** @nullable */
  buildingArea?: number | null;
  /** @nullable */
  basementArea?: number | null;
  /** @nullable */
  annexArea?: number | null;
  /** @nullable */
  floorsCount?: number | null;
  /** @nullable */
  coordinates?: string | null;
  /** @nullable */
  valuationMethod?: string | null;
  /** @nullable */
  marketValue?: number | null;
  /** @nullable */
  incomeValue?: number | null;
  /** @nullable */
  costValue?: number | null;
  /** @nullable */
  finalValue?: number | null;
  /** @nullable */
  pricePerMeter?: number | null;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  commercialRegNumber?: string | null;
  /** @nullable */
  notes?: string | null;
}

export type UpdateReportStatusBodyStatus =
  (typeof UpdateReportStatusBodyStatus)[keyof typeof UpdateReportStatusBodyStatus];

export const UpdateReportStatusBodyStatus = {
  pending: "pending",
  extracted: "extracted",
  reviewed: "reviewed",
  submitted: "submitted",
} as const;

export interface UpdateReportStatusBody {
  status: UpdateReportStatusBodyStatus;
}

export interface ReportStats {
  total: number;
  pending: number;
  extracted: number;
  reviewed: number;
  submitted: number;
}